Se agrego componente EstadisticasBoard
